pyproject.toml
[project]
name = "Personal cloud Database API"
version = "0.1.0"
description = "A personal cloud database API built with FastAPI and PostgreSQL."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyjwt>=2.10.1",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
